remove ordering and move article inside models to keep old code working
