+++ rackcity/models/__init__.py
@@ -1,4 +1,5 @@
-from it_instance import ITInstance
-from it_model import ITModel
-from rack import Rack
-from user import User
+from .it_instance import ITInstance
+from .it_model import ITModel
+from .rack import Rack
+from .user import User
+from .article import Article # from tutorial example
