+++ rackcity/models/it_model.py
@@ -5,13 +5,13 @@ class ITModel(models.Model):
     vendor = models.CharField(max_length=120)
     model_number = models.CharField(max_length=120)
     height = models.IntegerField()
+    comment = models.TextField()
     display_color = models.CharField(max_length=120, null=True, blank=True)
     num_ethernet_ports = models.IntegerField(null=True, blank=True)
     num_power_ports = models.IntegerField(null=True, blank=True)
     cpu = models.CharField(max_length=120, null=True, blank=True)
     memory_gb = models.IntegerField(null=True, blank=True)
     storage = models.CharField(max_length=120, null=True, blank=True)
-    comment = models.TextField()
 
     def save(self, *args, **kwargs): 
         super().save(*args, **kwargs)
@@ -20,7 +20,5 @@ class ITModel(models.Model):
         super().delete(*args, **kwargs)
 
     class Meta: 
-        ordering = [model_id]
-
         def __str__(self):
             return self.model_id
